+++ src/compress/zstd_lazy.rs
@@ -435,7 +435,7 @@ unsafe extern "C" fn ZSTD_hash4Ptr(
     mut ptr: *const std::ffi::c_void,
     mut h: u32,
 ) -> usize {
-    return ZSTD_hash4(MEM_readLE32(ptr), h, 0 as u32) as usize;
+    return ZSTD_hash4(MEM_readLE32(ptr), h, 0) as usize;
 }
 unsafe extern "C" fn ZSTD_hash4PtrS(
     mut ptr: *const std::ffi::c_void,
@@ -453,7 +453,7 @@ unsafe extern "C" fn ZSTD_hash5Ptr(
     mut p: *const std::ffi::c_void,
     mut h: u32,
 ) -> usize {
-    return ZSTD_hash5(MEM_readLE64(p), h, 0 as u64);
+    return ZSTD_hash5(MEM_readLE64(p), h, 0);
 }
 unsafe extern "C" fn ZSTD_hash5PtrS(
     mut p: *const std::ffi::c_void,
@@ -471,7 +471,7 @@ unsafe extern "C" fn ZSTD_hash6Ptr(
     mut p: *const std::ffi::c_void,
     mut h: u32,
 ) -> usize {
-    return ZSTD_hash6(MEM_readLE64(p), h, 0 as u64);
+    return ZSTD_hash6(MEM_readLE64(p), h, 0);
 }
 unsafe extern "C" fn ZSTD_hash6PtrS(
     mut p: *const std::ffi::c_void,
@@ -489,7 +489,7 @@ unsafe extern "C" fn ZSTD_hash7Ptr(
     mut p: *const std::ffi::c_void,
     mut h: u32,
 ) -> usize {
-    return ZSTD_hash7(MEM_readLE64(p), h, 0 as u64);
+    return ZSTD_hash7(MEM_readLE64(p), h, 0);
 }
 unsafe extern "C" fn ZSTD_hash7PtrS(
     mut p: *const std::ffi::c_void,
@@ -506,7 +506,7 @@ unsafe extern "C" fn ZSTD_hash8Ptr(
     mut p: *const std::ffi::c_void,
     mut h: u32,
 ) -> usize {
-    return ZSTD_hash8(MEM_readLE64(p), h, 0 as u64);
+    return ZSTD_hash8(MEM_readLE64(p), h, 0);
 }
 unsafe extern "C" fn ZSTD_hash8PtrS(
     mut p: *const std::ffi::c_void,
@@ -1878,7 +1878,7 @@ pub unsafe extern "C" fn ZSTD_row_update(
     let rowMask = ((1 as std::ffi::c_uint) << rowLog)
         .wrapping_sub(1);
     let mls = MIN!(ms -> cParams.minMatch, 6);
-    ZSTD_row_update_internal(ms, ip, mls, rowLog, rowMask, 0 as u32);
+    ZSTD_row_update_internal(ms, ip, mls, rowLog, rowMask, 0);
 }
 #[inline(always)]
 unsafe extern "C" fn ZSTD_row_matchMaskGroupWidth(rowEntries: u32) -> u32 {
